requirements.txt
catboost==1.2.2
DateTime==5.2
Flask==3.0.0
joblib==1.3.2
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.1
statsmodels==0.14.1
supervisor==4.2.5
Unidecode==1.3.7
urllib3==2.0.6
xgboost==2.0.0
scikit-learn==1.3.1